@@ -1,23 +1,25 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.1.2-3'
+    ext.kotlin_version = '1.3.72'
     repositories {
         jcenter()
+        maven { url 'https://jitpack.io' }
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.2'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$rootProject.ext.kotlin_version"
+        classpath 'com.android.tools.build:gradle:4.0.1'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
 }
-apply plugin: 'kotlin'
-
 allprojects {
     repositories {
         jcenter()
+        maven { url 'https://jitpack.io' }
+        google()
     }
 }
 
@@ -25,12 +27,10 @@ allprojects {
 //    delete rootProject.buildDir
 //}
 
-ext{
-    compileSdkVersion = 25
-    buildToolsVersion = "25.0.2"
-    supportLibVersion = "25.3.1"
-    kotlin_version="1.1.2-3"
-}
-repositories {
-    mavenCentral()
+ext {
+    minSdkVersion = 21
+    targetSdkVersion = 28
+    compileSdkVersion = 28
+    buildToolsVersion = "28.0.3"
+    supportLibVersion = "28.0.0"
 }
